hw: unify board definitions of bootstrap/core
Ref #3326
