+++ repos/base-hw/src/bootstrap/spec/arndale/board.h
@@ -14,24 +14,11 @@
 #ifndef _SRC__BOOTSTRAP__SPEC__ARNDALE__BOARD_H_
 #define _SRC__BOOTSTRAP__SPEC__ARNDALE__BOARD_H_
 
-#include <drivers/defs/arndale.h>
-#include <drivers/uart/exynos.h>
-
-#include <hw/spec/arm/cortex_a15.h>
+#include <hw/spec/arm/arndale_board.h>
 #include <hw/spec/arm/lpae.h>
 #include <spec/arm/cpu.h>
 #include <spec/arm/pic.h>
 
-namespace Board {
-
-	using namespace Arndale;
-	using Cpu_mmio = Hw::Cortex_a15_mmio<IRQ_CONTROLLER_BASE>;
-	using Serial = Genode::Exynos_uart;
-
-	enum {
-		UART_BASE  = UART_2_MMIO_BASE,
-		UART_CLOCK = UART_2_CLOCK,
-	};
-}
+namespace Board { using namespace Hw::Arndale_board; }
 
 #endif /* _SRC__BOOTSTRAP__SPEC__ARNDALE__BOARD_H_ */
