+++ repos/base-hw/src/bootstrap/spec/imx6q_sabrelite/board.h
@@ -14,29 +14,16 @@
 #ifndef _SRC__BOOTSTRAP__SPEC__IMX6Q_SABRELITE__BOARD_H_
 #define _SRC__BOOTSTRAP__SPEC__IMX6Q_SABRELITE__BOARD_H_
 
-#include <drivers/defs/imx6q_sabrelite.h>
-#include <drivers/uart/imx.h>
-#include <hw/spec/arm/cortex_a9.h>
-#include <hw/spec/arm/pl310.h>
-
+#include <hw/spec/arm/imx6q_sabrelite_board.h>
 #include <spec/arm/cortex_a9_actlr.h>
 #include <spec/arm/cortex_a9_page_table.h>
 #include <spec/arm/cpu.h>
 #include <spec/arm/pic.h>
 
 namespace Board {
-
-	using namespace Imx6q_sabrelite;
+	using namespace Hw::Imx6q_sabrelite_board;
 
 	struct L2_cache;
-	using Cpu_mmio = Hw::Cortex_a9_mmio<CORTEX_A9_PRIVATE_MEM_BASE>;
-	using Serial   = Genode::Imx_uart;
-
-	enum {
-		UART_BASE  = UART_2_MMIO_BASE,
-		UART_SIZE  = UART_2_MMIO_SIZE,
-		UART_CLOCK = 0, /* dummy value, not used */
-	};
 
 	static volatile unsigned long initial_values[][2] {
 		// (IOMUX Controller)
