+++ repos/base-hw/src/bootstrap/spec/imx7d_sabre/board.h
@@ -14,25 +14,13 @@
 #ifndef _SRC__BOOTSTRAP__SPEC__IMX7_SABRELITE__BOARD_H_
 #define _SRC__BOOTSTRAP__SPEC__IMX7_SABRELITE__BOARD_H_
 
-#include <drivers/defs/imx7d_sabre.h>
-#include <drivers/uart/imx.h>
-
-#include <hw/spec/arm/cortex_a15.h>
+#include <hw/spec/arm/imx7d_sabre_board.h>
 #include <hw/spec/arm/lpae.h>
 #include <spec/arm/cpu.h>
 #include <spec/arm/pic.h>
 
 namespace Board {
-
-	using namespace Imx7d_sabre;
-
-	using Cpu_mmio = Hw::Cortex_a15_mmio<IRQ_CONTROLLER_BASE>;
-	using Serial = Genode::Imx_uart;
-
-	enum {
-		UART_BASE  = UART_1_MMIO_BASE,
-		UART_CLOCK = 0, /* unsued value */
-	};
+	using namespace Hw::Imx7d_sabre_board;
 }
 
 #endif /* _SRC__BOOTSTRAP__SPEC__IMX&_SABRELITE__BOARD_H_ */
