+++ repos/base-hw/src/bootstrap/spec/nit6_solox/board.h
@@ -14,32 +14,18 @@
 #ifndef _SRC__BOOTSTRAP__SPEC__NIT6_SOLOX__BOARD_H_
 #define _SRC__BOOTSTRAP__SPEC__NIT6_SOLOX__BOARD_H_
 
-#include <drivers/defs/nit6_solox.h>
-#include <drivers/uart/imx.h>
-#include <hw/spec/arm/cortex_a9.h>
-#include <hw/spec/arm/pl310.h>
-
+#include <hw/spec/arm/nit6_solox_board.h>
 #include <spec/arm/cortex_a9_actlr.h>
 #include <spec/arm/cortex_a9_page_table.h>
 #include <spec/arm/cpu.h>
 #include <spec/arm/pic.h>
 
 namespace Board {
-
-	using namespace Nit6_solox;
+	using namespace Hw::Nit6_solox_board;
 
 	struct L2_cache;
-	using Cpu_mmio = Hw::Cortex_a9_mmio<CORTEX_A9_PRIVATE_MEM_BASE>;
-	using Serial   = Genode::Imx_uart;
-
-	enum {
-		UART_BASE  = UART_1_MMIO_BASE,
-		UART_SIZE  = UART_1_MMIO_SIZE,
-		UART_CLOCK = 0, /* dummy value, not used */
-	};
 
 	static volatile unsigned long initial_values[][2] {
-
 		// (IOMUX Controller)
 		{ 0x20E006C, 0x0},
 		{ 0x20E00CC, 0x10},
@@ -242,7 +228,6 @@ namespace Board {
 		{ 0x20C8100, 0x5258d0db}
 	};
 }
-
 
 struct Board::L2_cache : Hw::Pl310
 {
