+++ repos/base-hw/src/bootstrap/spec/pbxa9/board.h
@@ -14,10 +14,7 @@
 #ifndef _SRC__BOOTSTRAP__SPEC__PBXA9__BOARD_H_
 #define _SRC__BOOTSTRAP__SPEC__PBXA9__BOARD_H_
 
-#include <drivers/defs/pbxa9.h>
-#include <drivers/uart/pl011.h>
-#include <hw/spec/arm/cortex_a9.h>
-#include <hw/spec/arm/pl310.h>
+#include <hw/spec/arm/pbxa9_board.h>
 
 #include <spec/arm/cortex_a9_actlr.h>
 #include <spec/arm/cortex_a9_page_table.h>
@@ -25,17 +22,7 @@
 #include <spec/arm/pic.h>
 
 namespace Board {
-
-	using namespace Pbxa9;
-
-	using L2_cache = Hw::Pl310;
-	using Cpu_mmio = Hw::Cortex_a9_mmio<CORTEX_A9_PRIVATE_MEM_BASE>;
-	using Serial   = Genode::Pl011_uart;
-
-	enum {
-		UART_BASE  = PL011_0_MMIO_BASE,
-		UART_CLOCK = PL011_0_CLOCK,
-	};
+	using namespace Hw::Pbxa9_board;
 }
 
 #endif /* _SRC__BOOTSTRAP__SPEC__PBXA9__BOARD_H_ */
