+++ repos/base-hw/src/bootstrap/spec/riscv/board.h
@@ -14,22 +14,10 @@
 #ifndef _SRC__BOOTSTRAP__SPEC__RISCV__BOARD_H_
 #define _SRC__BOOTSTRAP__SPEC__RISCV__BOARD_H_
 
-#include <hw/spec/riscv/page_table.h>
-#include <hw/spec/riscv/uart.h>
-#include <drivers/defs/riscv.h>
-
-namespace Bootstrap {
-	struct Cpu {};
-	struct Pic {};
-}
+#include <hw/spec/riscv/board.h>
 
 namespace Board {
-
-	using namespace Riscv;
-
-	enum { UART_BASE, UART_CLOCK };
-	struct Serial : Hw::Riscv_uart {
-		Serial(Genode::addr_t, Genode::size_t, unsigned) {} };
+	using namespace Hw::Riscv_board;
 }
 
 template <typename E, unsigned B, unsigned S>
