+++ repos/base-hw/src/bootstrap/spec/rpi/board.h
@@ -14,19 +14,12 @@
 #ifndef _SRC__BOOTSTRAP__SPEC__RPI__BOARD_H_
 #define _SRC__BOOTSTRAP__SPEC__RPI__BOARD_H_
 
-#include <drivers/defs/rpi.h>
-#include <drivers/uart/pl011.h>
+#include <hw/spec/arm/rpi_board.h>
 #include <hw/spec/arm/page_table.h>
-
 #include <spec/arm/cpu.h>
 
 namespace Board {
-	using Serial   = Genode::Pl011_uart;
-
-	enum {
-		UART_BASE  = Rpi::PL011_0_MMIO_BASE,
-		UART_CLOCK = Rpi::PL011_0_CLOCK,
-	};
+	using namespace Hw::Rpi_board;
 }
 
 namespace Bootstrap { struct Pic {}; }
