+++ repos/base-hw/src/bootstrap/spec/x86_64/board.h
@@ -14,8 +14,7 @@
 #ifndef _SRC__BOOTSTRAP__SPEC__X86_64__BOARD_H_
 #define _SRC__BOOTSTRAP__SPEC__X86_64__BOARD_H_
 
-#include <drivers/uart/x86_pc.h>
-
+#include <hw/spec/x86_64/pc_board.h>
 #include <hw/spec/x86_64/page_table.h>
 #include <hw/spec/x86_64/cpu.h>
 #include <hw/spec/x86_64/x86_64.h>
@@ -26,14 +25,7 @@ namespace Bootstrap {
 }
 
 namespace Board {
-	struct Serial;
-	enum Dummies { UART_BASE, UART_CLOCK };
+	using namespace Hw::Pc_board;
 }
-
-
-struct Board::Serial : Genode::X86_uart
-{
-	Serial(Genode::addr_t, Genode::size_t, unsigned);
-};
 
 #endif /* _SRC__BOOTSTRAP__SPEC__X86_64__BOARD_H_ */
