+++ repos/base-hw/src/bootstrap/spec/zynq_qemu/board.h
@@ -14,25 +14,14 @@
 #ifndef _SRC__BOOTSTRAP__SPEC__ZYNQ__BOARD_H_
 #define _SRC__BOOTSTRAP__SPEC__ZYNQ__BOARD_H_
 
-#include <drivers/defs/zynq_qemu.h>
-#include <drivers/uart/xilinx.h>
-#include <hw/spec/arm/cortex_a9.h>
-#include <hw/spec/arm/pl310.h>
-
+#include <hw/spec/arm/zynq_qemu_board.h>
 #include <spec/arm/cortex_a9_actlr.h>
 #include <spec/arm/cortex_a9_page_table.h>
 #include <spec/arm/cpu.h>
 #include <spec/arm/pic.h>
 
 namespace Board {
-	using namespace Zynq_qemu;
-	using L2_cache = Hw::Pl310;
-	using Cpu_mmio = Hw::Cortex_a9_mmio<CORTEX_A9_PRIVATE_MEM_BASE>;
-	using Serial   = Genode::Xilinx_uart;
-
-	enum {
-		UART_BASE  = UART_0_MMIO_BASE,
-	};
+	using namespace Hw::Zynq_qemu_board;
 }
 
 #endif /* _SRC__BOOTSTRAP__SPEC__ZYNQ__BOARD_H_ */
