+++ repos/base-hw/src/core/spec/arndale/board.h
@@ -14,21 +14,10 @@
 #ifndef _CORE__SPEC__ARNDALE__BOARD_H_
 #define _CORE__SPEC__ARNDALE__BOARD_H_
 
-/* base includes */
-#include <drivers/defs/arndale.h>
-#include <drivers/uart/exynos.h>
-
-#include <hw/spec/arm/cortex_a15.h>
+#include <hw/spec/arm/arndale_board.h>
 
 namespace Board {
-	using namespace Arndale;
-	using Cpu_mmio = Hw::Cortex_a15_mmio<IRQ_CONTROLLER_BASE>;
-	using Serial = Genode::Exynos_uart;
-
-	enum {
-		UART_BASE  = UART_2_MMIO_BASE,
-		UART_CLOCK = UART_2_CLOCK,
-	};
+	using namespace Hw::Arndale_board;
 
 	static constexpr bool SMP = true;
 }
