+++ repos/base-hw/src/core/spec/imx6q_sabrelite/board.h
@@ -14,23 +14,13 @@
 #ifndef _CORE__SPEC__IMX6Q_SABRELITE__BOARD_H_
 #define _CORE__SPEC__IMX6Q_SABRELITE__BOARD_H_
 
-/* base includes */
-#include <drivers/defs/imx6q_sabrelite.h>
-#include <drivers/uart/imx.h>
-
-#include <hw/spec/arm/cortex_a9.h>
-#include <hw/spec/arm/pl310.h>
+#include <hw/spec/arm/imx6q_sabrelite_board.h>
 
 namespace Board {
-	using namespace Imx6q_sabrelite;
-	using Cpu_mmio = Hw::Cortex_a9_mmio<CORTEX_A9_PRIVATE_MEM_BASE>;
-	using L2_cache = Hw::Pl310;
-	using Serial   = Genode::Imx_uart;
 
-	enum {
-		UART_BASE  = UART_2_MMIO_BASE,
-		UART_CLOCK = 0, /* dummy value, not used */
-	};
+	using namespace Hw::Imx6q_sabrelite_board;
+
+	using L2_cache = Hw::Pl310;
 
 	static constexpr bool SMP = true;
