+++ repos/base-hw/src/core/spec/nit6_solox/board.h
@@ -14,23 +14,12 @@
 #ifndef _CORE__SPEC__NIT6_SOLOX__BOARD_H_
 #define _CORE__SPEC__NIT6_SOLOX__BOARD_H_
 
-/* base includes */
-#include <drivers/defs/nit6_solox.h>
-#include <drivers/uart/imx.h>
-
-#include <hw/spec/arm/cortex_a9.h>
-#include <hw/spec/arm/pl310.h>
+#include <hw/spec/arm/nit6_solox_board.h>
 
 namespace Board {
-	using namespace Nit6_solox;
-	using Cpu_mmio = Hw::Cortex_a9_mmio<CORTEX_A9_PRIVATE_MEM_BASE>;
-	using L2_cache = Hw::Pl310;
-	using Serial   = Genode::Imx_uart;
+	using namespace Hw::Nit6_solox_board;
 
-	enum {
-		UART_BASE  = UART_1_MMIO_BASE,
-		UART_CLOCK = 0, /* dummy value, not used */
-	};
+	using L2_cache = Hw::Pl310;
 
 	static constexpr bool SMP = true;
