+++ repos/base-hw/src/core/spec/panda/board.h
@@ -15,25 +15,11 @@
 #ifndef _CORE__SPEC__PANDA__BOARD_H_
 #define _CORE__SPEC__PANDA__BOARD_H_
 
-/* base includes */
-#include <drivers/defs/panda.h>
-#include <drivers/uart/tl16c750.h>
-
-#include <hw/spec/arm/cortex_a9.h>
-#include <hw/spec/arm/pl310.h>
-#include <hw/spec/arm/panda_trustzone_firmware.h>
+#include <hw/spec/arm/panda_board.h>
 
 namespace Board {
-
-	using namespace Panda;
-	using Cpu_mmio = Hw::Cortex_a9_mmio<CORTEX_A9_PRIVATE_MEM_BASE>;
-	using Serial   = Genode::Tl16c750_uart;
-
-	enum {
-		UART_BASE  = TL16C750_3_MMIO_BASE,
-		UART_CLOCK = TL16C750_CLOCK,
-	};
-
+	using namespace Hw::Panda_board;
+	
 	static constexpr bool SMP = true;
 
 	class L2_cache : public Hw::Pl310
