+++ repos/base-hw/src/core/spec/pbxa9/board.h
@@ -14,24 +14,10 @@
 #ifndef _CORE__SPEC__PBXA9__BOARD_H_
 #define _CORE__SPEC__PBXA9__BOARD_H_
 
-/* base includes */
-#include <drivers/defs/pbxa9.h>
-#include <drivers/uart/pl011.h>
-
-#include <hw/spec/arm/cortex_a9.h>
-#include <hw/spec/arm/pl310.h>
+#include <hw/spec/arm/pbxa9_board.h>
 
 namespace Board {
-	using namespace Pbxa9;
-
-	using Cpu_mmio = Hw::Cortex_a9_mmio<CORTEX_A9_PRIVATE_MEM_BASE>;
-	using L2_cache = Hw::Pl310;
-	using Serial   = Genode::Pl011_uart;
-
-	enum {
-		UART_BASE  = PL011_0_MMIO_BASE,
-		UART_CLOCK = PL011_0_CLOCK,
-	};
+	using namespace Hw::Pbxa9_board;
 
 	static constexpr bool SMP = true;
