+++ repos/base-hw/src/core/spec/rpi/board.h
@@ -15,18 +15,10 @@
 #ifndef _CORE__SPEC__RPI__BOARD_H_
 #define _CORE__SPEC__RPI__BOARD_H_
 
-#include <drivers/defs/rpi.h>
-#include <drivers/uart/pl011.h>
+#include <hw/spec/arm/rpi_board.h>
 
 namespace Board {
-	using namespace Rpi;
-
-	using Serial   = Genode::Pl011_uart;
-
-	enum {
-		UART_BASE  = Rpi::PL011_0_MMIO_BASE,
-		UART_CLOCK = Rpi::PL011_0_CLOCK,
-	};
+	using namespace Hw::Rpi_board;
 
 	static constexpr bool SMP = false;
 };
