+++ repos/base-hw/src/core/spec/wand_quad/board.h
@@ -15,24 +15,13 @@
 #ifndef _CORE__SPEC__WAND_QUAD__BOARD_H_
 #define _CORE__SPEC__WAND_QUAD__BOARD_H_
 
-/* base includes */
-#include <drivers/defs/wand_quad.h>
-#include <drivers/uart/imx.h>
-
-#include <hw/spec/arm/cortex_a9.h>
-#include <hw/spec/arm/pl310.h>
+#include <hw/spec/arm/wand_quad_board.h>
 
 namespace Board {
-	using namespace Wand_quad;
-	using Cpu_mmio = Hw::Cortex_a9_mmio<CORTEX_A9_PRIVATE_MEM_BASE>;
+	using namespace Hw::Wand_quad_board;
+
 	using L2_cache = Hw::Pl310;
-	using Serial   = Genode::Imx_uart;
-
-	enum {
-		UART_BASE  = UART_1_MMIO_BASE,
-		UART_CLOCK = 0, /* dummy value, not used */
-	};
-
+	
 	static constexpr bool SMP = true;
 
 	L2_cache & l2_cache();
