+++ repos/base-hw/src/include/hw/spec/arm/panda_board.h
@@ -1,22 +1,34 @@
 /*
- * \brief  Pandaboard's TrustZone firmware frontend
- * \author Stefan Kalkowski
- * \date   2017-02-02
+ * \brief   Pandaboard specific definitions
+ * \author  Stefan Kalkowski
+ * \date    2019-05-16
  */
 
 /*
- * Copyright (C) 2017 Genode Labs GmbH
+ * Copyright (C) 2019 Genode Labs GmbH
  *
  * This file is part of the Genode OS framework, which is distributed
  * under the terms of the GNU Affero General Public License version 3.
  */
 
-#ifndef _SRC__LIB__HW__SPEC__ARM__PANDA_TRUSTZONE_FIRMWARE_H_
-#define _SRC__LIB__HW__SPEC__ARM__PANDA_TRUSTZONE_FIRMWARE_H_
+#ifndef _SRC__INCLUDE__HW__SPEC__ARM__PANDA_BOARD_H_
+#define _SRC__INCLUDE__HW__SPEC__ARM__PANDA_BOARD_H_
 
 #include <base/stdint.h>
+#include <drivers/defs/panda.h>
+#include <drivers/uart/tl16c750.h>
+#include <hw/spec/arm/cortex_a9.h>
+#include <hw/spec/arm/pl310.h>
 
-namespace Hw {
+namespace Hw::Panda_board {
+	using namespace Panda;
+	using Cpu_mmio = Hw::Cortex_a9_mmio<CORTEX_A9_PRIVATE_MEM_BASE>;
+	using Serial   = Genode::Tl16c750_uart;
+
+	enum {
+		UART_BASE  = TL16C750_3_MMIO_BASE,
+		UART_CLOCK = TL16C750_CLOCK,
+	};
 
 	enum Panda_firmware_opcodes {
 		CPU_ACTLR_SMP_BIT_RAISE =  0x25,
@@ -38,4 +50,4 @@ namespace Hw {
 	}
 }
 
-#endif /* _SRC__LIB__HW__SPEC__ARM__PANDA_TRUSTZONE_FIRMWARE_H_ */
+#endif /* _SRC__INCLUDE__HW__SPEC__ARM__PANDA_BOARD_H_ */
